MGDOBR-1215: Manager: Remove manager-core test-jar usage
