+++ manager-parent/manager-core/src/test/java/com/redhat/service/smartevents/manager/core/persistence/EnumTypeManagedResourceStatusTest.java
@@ -1,16 +1,84 @@
 package com.redhat.service.smartevents.manager.core.persistence;
 
+import java.util.Properties;
+
+import org.hibernate.type.EnumType;
+import org.hibernate.type.descriptor.java.EnumJavaTypeDescriptor;
+import org.hibernate.type.descriptor.java.spi.JavaTypeDescriptorRegistry;
+import org.hibernate.type.spi.TypeConfiguration;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
 import com.redhat.service.smartevents.infra.core.models.ManagedResourceStatus;
 
-public class EnumTypeManagedResourceStatusTest extends EnumTypeBaseTest<ManagedResourceStatus, EnumTypeManagedResourceStatus> {
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
 
-    @Override
-    protected EnumTypeManagedResourceStatus getType() {
-        return new EnumTypeManagedResourceStatus();
+@ExtendWith(MockitoExtension.class)
+public class EnumTypeManagedResourceStatusTest {
+
+    @Mock
+    TypeConfiguration typeConfiguration;
+
+    @Mock
+    JavaTypeDescriptorRegistry javaTypeDescriptorRegistry;
+
+    @Mock
+    @SuppressWarnings("raw")
+    EnumJavaTypeDescriptor enumJavaTypeDescriptor;
+
+    private EnumTypeManagedResourceStatus type;
+
+    private enum TestEnum {
     }
 
-    @Override
-    protected Class<ManagedResourceStatus> getEnumType() {
-        return ManagedResourceStatus.class;
+    @BeforeEach
+    @SuppressWarnings("unchecked")
+    public void setup() {
+        this.type = new EnumTypeManagedResourceStatus();
+        this.type.setTypeConfiguration(typeConfiguration);
+
+        when(typeConfiguration.getJavaTypeDescriptorRegistry()).thenReturn(javaTypeDescriptorRegistry);
+        when(javaTypeDescriptorRegistry.getDescriptor(any())).thenReturn(enumJavaTypeDescriptor);
     }
+
+    @Test
+    public void testSetParameterValuesOverrideWhenNull() {
+        type.setParameterValues(null);
+
+        assertDefaultInstantiation();
+    }
+
+    @Test
+    public void testSetParameterValuesOverrideWhenEmpty() {
+        Properties parameters = new Properties();
+
+        type.setParameterValues(parameters);
+
+        assertDefaultInstantiation();
+        assertThat(parameters.isEmpty()).isTrue();
+    }
+
+    @Test
+    public void testSetParameterValuesOverrideWhenPrePopulated() {
+        Properties parameters = new Properties();
+        parameters.setProperty(EnumType.ENUM, TestEnum.class.getName());
+        parameters.setProperty(EnumType.NAMED, Boolean.FALSE.toString());
+
+        type.setParameterValues(parameters);
+
+        // When the properties are already set they take precedence over our workaround
+        assertThat(parameters.get(EnumType.ENUM)).isEqualTo(TestEnum.class.getName());
+        assertThat(parameters.get(EnumType.NAMED)).isEqualTo(Boolean.FALSE.toString());
+    }
+
+    private void assertDefaultInstantiation() {
+        assertThat(type.isOrdinal()).isFalse();
+        assertThat(type.returnedClass()).isEqualTo(ManagedResourceStatus.class);
+    }
+
 }
