+++ manager-parent/manager-v1/src/main/resources/db/migration/0.1/V0.1.5__create_shards.sql
@@ -1,5 +1,3 @@
-DROP TABLE IF EXISTS SHARD;
-
 create table SHARD
 (
     id   varchar(255) NOT NULL PRIMARY KEY,
