+++ manager-parent/manager-v1/src/test/java/com/redhat/service/smartevents/manager/v1/mocks/DefaultJsonWebToken.java
@@ -1,0 +1,28 @@
+package com.redhat.service.smartevents.manager.v1.mocks;
+
+import java.util.Set;
+
+import javax.enterprise.context.RequestScoped;
+
+import org.eclipse.microprofile.jwt.JsonWebToken;
+
+import com.redhat.service.smartevents.manager.v1.TestConstants;
+
+@RequestScoped
+public class DefaultJsonWebToken implements JsonWebToken {
+
+    @Override
+    public String getName() {
+        return TestConstants.SHARD_ID;
+    }
+
+    @Override
+    public Set<String> getClaimNames() {
+        return null;
+    }
+
+    @Override
+    public <T> T getClaim(String s) {
+        return null;
+    }
+}
