+++ manager-parent/manager-v1/src/test/resources/application.properties
@@ -25,10 +25,20 @@ event-bridge.kafka.client.secret=client-secret
 event-bridge.kafka.security.protocol=PLAINTEXT
 event-bridge.kafka.sasl.mechanism=PLAIN
 
+# RHOAS Client
+event-bridge.rhoas.instance-api.host=http://test-instance-api.redhat.com
+event-bridge.rhoas.mgmt-api.host=http://test-management-api.redhat.com
+# RHOAS Client is unused by -core so Quarkus removes it from Arc unless we flag it as unremovable.
+quarkus.arc.unremovable-types=com.redhat.service.smartevents.rhoas.RhoasClient
+
 # Quartz Scheduler configuration for tests
 event-bridge.resources.workers.job-retry-interval-seconds=1
 
 # Disable error processing scheduler
 event-bridge.processing-errors.cleanup.schedule=0 0 0 1 1 ? 2099
+
+# LOGGING
+# Use unstructured logging in test
+quarkus.log.console.json=false
 
 event-bridge.account-management-service.quota.config={"myOrg": {"bridges_quota": 100, "processors_quota": 100}}
